+++ src/pages/@owner/food-truck-form/@section/menu-section/MenuInfo.tsx
@@ -3,7 +3,7 @@ import { useFormContext } from 'react-hook-form';
 import { ROUTES } from '@router/constant/routes';
 import FormLayout from '@pages/@owner/food-truck-form/components/FormLayout';
 import PageSwitchButton from '@pages/@owner/food-truck-form/components/PageSwitchButton';
-import type { FoodTruckFormData } from '@pages/@owner/food-truck-form/hooks/use-food-truck-form';
+import type { FoodTruckFormData } from '@pages/@owner/food-truck-form/utils/use-food-truck-form';
 import { FOOD_TRUCK_ERROR_MESSAGE } from '@pages/@owner/food-truck-form/constants/food-truck';
 import { getNavigateState } from '@pages/@owner/food-truck-form/utils/navigate-state';
