+++ src/pages/@owner/food-truck-form/hooks/use-menu-info.ts
@@ -1,5 +1,5 @@
 import { useFormContext } from 'react-hook-form';
-import type { FoodTruckFormData } from '@pages/@owner/food-truck-form/hooks/use-food-truck-form';
+import type { FoodTruckFormData } from '@pages/@owner/food-truck-form/utils/use-food-truck-form';
 
 export const useMenuInfo = () => {
   const { setValue } = useFormContext<FoodTruckFormData>();
