+++ src/pages/@owner/food-truck-form/utils/navigate-state.tsx
@@ -1,4 +1,4 @@
-import type { FoodTruckFormData } from '@pages/@owner/food-truck-form/hooks/use-food-truck-form';
+import type { FoodTruckFormData } from '@pages/@owner/food-truck-form/utils/use-food-truck-form';
 
 export const getNavigateState = (formData: FoodTruckFormData) => {
   return {
